pyproject.toml
[project]
name = "news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "youtube-transcript-api>=1.2.4",
]
